Fix order detail refund

+++ app/Filament/Resources/OrderResource/RelationManagers/OrderDetailsRelationManager.php
@@ -15,6 +15,7 @@ use Filament\Forms\Form;
 use Filament\Tables\Table;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Actions;
+use Filament\Notifications\Notification;
 
 class OrderDetailsRelationManager extends RelationManager
 {
@@ -105,32 +106,35 @@ class OrderDetailsRelationManager extends RelationManager
                             ]),
                     ])
                     ->modalSubmitActionLabel('Approve')
-                    ->extraModalFooterActions([
-                        Actions\Action::make('reject')
-                            ->color('danger')
-                            ->requiresConfirmation()
-                            ->action(function ($action, $record) {
-                                $record->orderDetailRefund->delete();
-                                $action->success();
-                            })
-                            ->successNotificationTitle('Refund Request rejected')
-                            ->successNotification(function ($notification, $record) {
-                                $this->getOwnerRecord()
-                                    ->user
-                                    ->notify(new SendUserOrderRefundRejected(
-                                        $this->getOwnerRecord()->order_column,
-                                        \implode(' - ', [
-                                            $record->discount->brand->name,
-                                            $record->discount->name,
-                                        ]
-                                    )));
-                                return $notification;
-                            }),
+                    ->extraModalFooterActions(fn ($action) => [
+                        $action->makeModalSubmitAction('reject', arguments: ['reject' => true])
+                            ->color('danger'),
                     ])
-                    ->action(function ($action, $record) {
+                    ->action(function ($action, $record, $arguments) {
+                        if ($arguments['reject'] ?? false) {
+                            $record->orderDetailRefund->delete();
+
+                            $this->getOwnerRecord()
+                                ->user
+                                ->notify(new SendUserOrderRefundRejected(
+                                    $this->getOwnerRecord()->order_column,
+                                    \implode(' - ', [
+                                        $record->discount->brand->name,
+                                        $record->discount->name,
+                                    ]
+                                )));
+
+                            Notification::make()
+                                ->success()
+                                ->title('Refund request rejected')
+                                ->send();
+
+                            return;
+                        }
+
                         $record->quantity = $record->orderDetailRefund->quantity;
 
-                        (new CreateCcRefund($record->cardknox_refnum, $record->total / 100))->send();
+                        (new CreateCcRefund($record->order->cardknox_refnum, $record->total / 100))->send();
 
                         $record->orderDetailRefund->update([
                             'approved_at' => \now(),
